% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BUBBLE2}
\alias{BUBBLE2}
\title{Data for the Function}
\format{
A data frame with 18 rows and 5 variables:
\describe{
  \item{Density}{Density of the Bubble}
  \item{Diameter}{Diameter of the Bubble}
  \item{HeatFlux}{Changes in heat}
  \item{Label}{Labels of each data}
  \item{MassFlux}{Fluxuation in the Mass}
  ...
}
}
\source{
\url{https://www.routledge.com/Statistics-for-Engineering-and-the-Sciences/Mendenhall-Sincich/p/book/9781498728850}
}
\usage{
BUBBLE2
}
\description{
A data set containing information about bubble behavior in subcooled flow boiling.
}
\keyword{datasets}
