% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot.R
\name{myboot}
\alias{myboot}
\title{Bootstraping}
\usage{
myboot(df, model = "model1", iter = 1, alpha = 0.05)
}
\arguments{
\item{df}{A data Frame}

\item{model}{A string indicating model1 or model2 (diameter or density)}

\item{iter}{Iterations of the sampling test}

\item{alpha}{Alpha value to create a confidence interval}
}
\value{
The function will make a histogram plot of each \eqn{\beta_i} from
   the sampling process. An invisible list will also be returned that contains
   the following:
   \itemize{
    \item{\code{beta_iters}}{ - A data frame of each \eqn{\beta} calculated from the iterations of sampling the `BUBBLE2` data}
    \item{\code{CI_beta}}{ - The confidence interval of the original point estimates of \eqn{\beta}}
    \item{\code{beta_pts}}{ - The point estimates of \eqn{\beta}}
    \item{\code{iter}}{ - The passed in value for the number of iterations}
    \item{\code{model}}{ - Chosen model for the function}
    \item{\code{alpha}}{ - Chosen alpha for the function}
   }
}
\description{
Bootstraping
}
\examples{
dataf <- data.frame(BUBBLE2)

temp <- myboot(dataf, model = "model1", iter = 200, alpha = 0.05)
temp

temp <- myboot(dataf, model = "model2", iter = 1000, alpha = 0.025)
temp
}
